Added mmap_lat parsing args

+++ engines/mmap_lat.h
@@ -7,9 +7,9 @@ void mmap_lat_engine(Mapping *, Arguments *, Results *);
 void mmap_lat_check_args(Arguments *);
 int mmap_lat_prep_file(Arguments args);
 
-double mmap_lat_do_mmap(Mapping *, Arguments, int);
+uint64_t mmap_lat_do_mmap(Mapping *, Arguments, int);
 void mmap_lat_do_unmap(Mapping *);
-double mmap_lat_do_mmap_anon(Mapping *mapping, uint64_t fsize, int fd);
+uint64_t mmap_lat_do_mmap_anon(Mapping *mapping, uint64_t fsize, int fd);
 void mmap_lat_cleanup_file(Mapping *mapping, int fd);
 
 #endif
